requirements.txt
art==6.1
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
DateTime==5.4
Django==5.0.3
flatbuffers==23.5.26
humanfriendly==10.0
idna==3.6
imageio==2.34.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy_loader==0.3
llvmlite==0.42.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
onnxruntime==1.17.1
opencv-python-headless==4.9.0.80
packaging==23.2
pathlib==1.0.1
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
progress==1.6
protobuf==4.25.3
psycopg2-binary==2.9.9
PyMatting==1.1.12
pyreadline3==3.4.1
pyTelegramBotAPI==4.16.1
pytz==2024.1
referencing==0.33.0
rembg==2.0.55
requests==2.31.0
rpds-py==0.18.0
scikit-image==0.22.0
scipy==1.12.0
setuptools==69.1.1
sqlparse==0.4.4
sympy==1.12
telebot==0.0.5
tifffile==2024.2.12
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
zope.interface==6.2
